localllmtest/requirements.txt
llama-cpp-python
fastapi
uvicorn
huggingface_hub
Pillow